deleting all strategies, most commands
