+++ editor/src/components/canvas/canvas-types.ts
@@ -24,10 +24,6 @@ import {
   LayoutFlexElementProp,
   LayoutTargetableProp,
 } from '../../core/layout/layout-helpers-new'
-import {
-  CanvasInteractionSession,
-  SelectModeCanvasSessionProps,
-} from './canvas-strategies/canvas-strategy-types'
 import { InteractionState, InteractionStateWithoutMetadata } from './interactions/interaction-state'
 
 export const CanvasContainerID = 'canvas-container'
