+++ editor/src/templates/editor-canvas.tsx
@@ -92,7 +92,6 @@ import {
   updateGlobalPositions,
 } from '../utils/global-positions'
 import { last, reverse } from '../core/shared/array-utils'
-import { updateSelectModeCanvasSessionDragVector } from '../components/canvas/canvas-strategies/canvas-strategy-types'
 import { updateInteractionViaMouse } from '../components/canvas/interactions/interaction-state'
 
 const webFrame = PROBABLY_ELECTRON ? requireElectron().webFrame : null
